+++ src/interpreter/builtins/mod.rs
@@ -66,6 +66,7 @@ lazy_static! {
         array_methods.insert("length".to_string(), ITER_LEN.clone());
         array_methods.insert("map".to_string(), iterable::ITER_MAP.clone());
         array_methods.insert("filter".to_string(), iterable::ITER_FILTER.clone());
+        array_methods.insert("reduce".to_string(), iterable::ITER_REDUCE.clone());
         array_methods.insert("concat".to_string(), concat::CONCAT.clone());
         array_methods.insert("format".to_string(), format::NON_NUM_FORMAT.clone());
         m.insert(NonParametricType::Array, array_methods);